Move system generation complexity to generators instead of new
