+++ src/generators/mod.rs
@@ -2,7 +2,7 @@ use rand::{Rng, SeedableRng};
 use std::time::Instant;
 use rayon::prelude::*;
 use rand::isaac::IsaacRng;
-use statrs::distribution::{Distribution, Uniform, Normal};
+use statrs::distribution::{Distribution, Gamma, Uniform, Normal};
 use nalgebra::geometry::Point3 as Point;
 use std::sync::{Arc, Mutex};
 
@@ -11,9 +11,14 @@ pub mod stars;
 pub mod planets;
 
 use resources::{fetch_resource, StarTypesResource, PlanetTypesResource, AstronomicalNamesResource};
-use astronomicals::Galaxy;
+use astronomicals::{Galaxy, hash};
 use astronomicals::system::System;
 use game_config::GameConfig;
+use generators::stars::StarGen;
+use generators::names::NameGen;
+use generators::planets::PlanetGen;
+use astronomicals::star::Star;
+use astronomicals::planet::Planet;
 
 /// A generator that can be explicitly seeded in order to the produce the same
 /// stream of psuedo randomness each time
@@ -106,7 +111,7 @@ pub fn generate_galaxy(config: &GameConfig) -> Galaxy {
 
             // Generate systems
             for _ in 0..cluster_size {
-                cluster_systems.push(System::new(
+                cluster_systems.push(generate_system(
                     Point::new(
                         norm_x.sample::<IsaacRng>(&mut rng),
                         norm_y.sample::<IsaacRng>(&mut rng),
@@ -139,4 +144,51 @@ pub fn generate_galaxy(config: &GameConfig) -> Galaxy {
         systems.iter().take(5).collect::<Vec<_>>()
     );
     Galaxy::new(systems)
+}
+
+// Generate a new star system using the given generators and a location as seed
+pub fn generate_system(
+    location: Point<f64>,
+    name_gen: Arc<Mutex<NameGen>>,
+    star_gen: &StarGen,
+    planet_gen: &PlanetGen,
+) -> System {
+    // Calculate hash
+    let hash = hash(location);
+    let seed: &[_] = &[hash as u32];
+    let mut rng: IsaacRng = SeedableRng::from_seed(seed);
+
+    let star = star_gen.generate(&mut rng).unwrap();
+
+    // Unwrap and lock name generator as it is mutated by generation
+    let mut name_gen_unwraped = name_gen.lock().unwrap();
+
+    // TODO: Replace constant in config
+    let num_planets = Gamma::new(1., 0.5)
+        .unwrap()
+        .sample::<IsaacRng>(&mut rng)
+        .round() as u32;
+    let mut satelites: Vec<Planet> = (0..num_planets)
+        .map(|_| planet_gen.generate(&mut rng).unwrap())
+        .collect();
+
+    // Fallback to "Unnamed" for names
+    for planet in &mut satelites {
+        planet.set_name(name_gen_unwraped.generate().unwrap_or(
+            String::from("Unnamed"),
+        ));
+        planet.set_surface_temperature(&star);
+    }
+
+    // System name is the same as one random planet
+    let name = match rng.choose(&satelites) {
+        Some(planet) => planet.name.clone(),
+        None => {
+            name_gen_unwraped.generate().unwrap_or(
+                String::from("Unnamed"),
+            )
+        }
+    } + " System";
+
+    System::new(location, name, star, satelites)
 }
